Separate Core and IO libraries
