+++ src/Core/Camera/PinholeCameraIntrinsic.h
@@ -27,49 +27,49 @@
 #pragma once
 
 #include <Eigen/Core>
-#include <IO/ClassIO/IJsonConvertible.h>
+#include <Core/Utility/IJsonConvertible.h>
 
 namespace three {
 
 class PinholeCameraIntrinsic : public IJsonConvertible
 {
 public:
-	static const PinholeCameraIntrinsic PrimeSenseDefault;
+   static const PinholeCameraIntrinsic GetPrimeSenseDefault();
 
 public:
-	PinholeCameraIntrinsic();
-	PinholeCameraIntrinsic(int width, int height, double fx, double fy,
-			double cx, double cy);
-	~PinholeCameraIntrinsic() override;
+    PinholeCameraIntrinsic();
+    PinholeCameraIntrinsic(int width, int height, double fx, double fy,
+        double cx, double cy);
+    ~PinholeCameraIntrinsic() override;
 
 public:
-	void SetIntrinsics(int width, int height, double fx, double fy, double cx,
-			double cy) {
-		width_ = width; height_ = height;
-		intrinsic_matrix_.setIdentity();
-		intrinsic_matrix_(0, 0) = fx; intrinsic_matrix_(1, 1) = fy;
-		intrinsic_matrix_(0, 2) = cx; intrinsic_matrix_(1, 2) = cy;
-	}
+    void SetIntrinsics(int width, int height, double fx, double fy, double cx,
+            double cy) {
+        width_ = width; height_ = height;
+        intrinsic_matrix_.setIdentity();
+        intrinsic_matrix_(0, 0) = fx; intrinsic_matrix_(1, 1) = fy;
+        intrinsic_matrix_(0, 2) = cx; intrinsic_matrix_(1, 2) = cy;
+    }
 
-	std::pair<double, double> GetFocalLength() const {
-		return std::make_pair(intrinsic_matrix_(0, 0), intrinsic_matrix_(1, 1));
-	}
+    std::pair<double, double> GetFocalLength() const {
+        return std::make_pair(intrinsic_matrix_(0, 0), intrinsic_matrix_(1, 1));
+    }
 
-	std::pair<double, double> GetPrincipalPoint() const {
-		return std::make_pair(intrinsic_matrix_(0, 2), intrinsic_matrix_(1, 2));
-	}
+    std::pair<double, double> GetPrincipalPoint() const {
+        return std::make_pair(intrinsic_matrix_(0, 2), intrinsic_matrix_(1, 2));
+    }
 
-	double GetSkew() const { return intrinsic_matrix_(0, 1); }
+    double GetSkew() const { return intrinsic_matrix_(0, 1); }
 
-	bool IsValid() const { return (width_ > 0 && height_ > 0); }
+    bool IsValid() const { return (width_ > 0 && height_ > 0); }
 
-	bool ConvertToJsonValue(Json::Value &value) const override;
-	bool ConvertFromJsonValue(const Json::Value &value) override;
+    bool ConvertToJsonValue(Json::Value &value) const override;
+    bool ConvertFromJsonValue(const Json::Value &value) override;
 
 public:
-	int width_ = -1;
-	int height_ = -1;
-	Eigen::Matrix3d intrinsic_matrix_;
+    int width_ = -1;
+    int height_ = -1;
+    Eigen::Matrix3d intrinsic_matrix_;
 };
 
-}	// namespace three
+}   // namespace three
