+++ src/Core/Geometry/PointCloudFactory.cpp
@@ -31,7 +31,6 @@
 #include <Core/Geometry/Image.h>
 #include <Core/Geometry/RGBDImage.h>
 #include <Core/Camera/PinholeCameraIntrinsic.h>
-#include <IO/ClassIO/PointCloudIO.h>
 
 namespace three{
 
@@ -97,14 +96,6 @@ std::shared_ptr<PointCloud> CreatePointCloudFromRGBDImageT(
 }
 
 }	// unnamed namespace
-
-std::shared_ptr<PointCloud> CreatePointCloudFromFile(
-		const std::string &filename)
-{
-	auto pointcloud = std::make_shared<PointCloud>();
-	ReadPointCloud(filename, *pointcloud);
-	return pointcloud;
-}
 
 std::shared_ptr<PointCloud> CreatePointCloudFromDepthImage(
 		const Image &depth, const PinholeCameraIntrinsic &intrinsic,
