+++ src/Core/Registration/PoseGraphFactory.cpp
@@ -26,16 +26,7 @@
 
 #include "PoseGraph.h"
 
-#include <IO/ClassIO/PoseGraphIO.h>
-
 namespace three{
 
-std::shared_ptr<PoseGraph> CreatePoseGraphFromFile(
-		const std::string &filename)
-{
-	auto pose_graph = std::make_shared<PoseGraph>();
-	ReadPoseGraph(filename, *pose_graph);
-	return pose_graph;
-}
 
 }	// namespace three
