+++ src/IO/ClassIO/PointCloudIO.cpp
@@ -58,6 +58,14 @@ static const std::unordered_map<std::string,
 		};
 }	// unnamed namespace
 
+std::shared_ptr<PointCloud> CreatePointCloudFromFile(
+    const std::string &filename)
+{
+    auto pointcloud = std::make_shared<PointCloud>();
+    ReadPointCloud(filename, *pointcloud);
+    return pointcloud;
+}
+
 bool ReadPointCloud(const std::string &filename, PointCloud &pointcloud)
 {
 	std::string filename_ext =
