+++ src/IO/ClassIO/PoseGraphIO.h
@@ -31,6 +31,12 @@
 
 namespace three {
 
+/// Factory function to create a PoseGraph from a file
+/// (PinholeCameraTrajectoryFactory.cpp)
+/// Return an empty PinholeCameraTrajectory if fail to read the file.
+std::shared_ptr<PoseGraph> CreatePoseGraphFromFile(
+    const std::string &filename);
+
 /// The general entrance for reading a PoseGraph from a file.
 /// The function calls read functions based on the extension name of filename.
 /// \return return true if the read function is successful, false otherwise.
