+++ src/Python/Tutorial/Advanced/camera_trajectory.py
@@ -8,8 +8,8 @@ from py3d import *
 if __name__ == "__main__":
 
 	print("Testing camera in py3d ...")
-	print(PinholeCameraIntrinsic.prime_sense_default)
-	print(PinholeCameraIntrinsic.prime_sense_default.intrinsic_matrix)
+	print(PinholeCameraIntrinsic.get_prime_sense_default())
+	print(PinholeCameraIntrinsic.get_prime_sense_default().intrinsic_matrix)
 	print(PinholeCameraIntrinsic())
 	x = PinholeCameraIntrinsic(640, 480, 525, 525, 320, 240)
 	print(x)
