+++ src/Python/Tutorial/Advanced/rgbd_integration.py
@@ -17,7 +17,7 @@ if __name__ == "__main__":
         depth = read_image("../../TestData/RGBD/depth/{:05d}.png".format(i))
         rgbd = create_rgbd_image_from_color_and_depth(color, depth,
 				depth_trunc = 4.0, convert_rgb_to_intensity = False)
-        volume.integrate(rgbd, PinholeCameraIntrinsic.prime_sense_default,
+        volume.integrate(rgbd, PinholeCameraIntrinsic.get_prime_sense_default(),
 				np.linalg.inv(camera_poses[i].pose))
 
     print("Extract a triangle mesh from the volume and visualize it.")
