+++ src/Python/Tutorial/ReconstructionSystem/integrate_scene.py
@@ -57,5 +57,5 @@ if __name__ == "__main__":
 	if args.path_intrinsic:
 		intrinsic = read_pinhole_camera_intrinsic(args.path_intrinsic)
 	else:
-		intrinsic = PinholeCameraIntrinsic.prime_sense_default
+		intrinsic = PinholeCameraIntrinsic.get_prime_sense_default()
 	scalable_integrate_rgb_frames(args.path_dataset, intrinsic)
