+++ src/Python/Tutorial/ReconstructionSystem/make_fragments.py
@@ -117,7 +117,7 @@ def process_fragments(path_dataset, path_intrinsic):
 	if path_intrinsic:
 		intrinsic = read_pinhole_camera_intrinsic(path_intrinsic)
 	else:
-		intrinsic = PinholeCameraIntrinsic.prime_sense_default
+		intrinsic = PinholeCameraIntrinsic.get_prime_sense_default()
 
 	make_folder(path_dataset + folder_fragment)
 	[color_files, depth_files] = get_rgbd_file_lists(path_dataset)
