+++ src/Python/Tutorial/ReconstructionSystem/test_pairwise_rgbd_alignment.py
@@ -65,5 +65,5 @@ if __name__ == "__main__":
 	if args.path_intrinsic:
 		intrinsic = read_pinhole_camera_intrinsic(args.path_intrinsic)
 	else:
-		intrinsic = PinholeCameraIntrinsic.prime_sense_default
+		intrinsic = PinholeCameraIntrinsic.get_prime_sense_default()
 	test_single_pair(args.source_id, args.target_id, intrinsic, with_opencv)
