+++ src/Visualization/CMakeLists.txt
@@ -35,7 +35,7 @@ source_group("Header Files\\Visualizer" FILES ${VISUALIZATION_VISUALIZER_HEADER_
 file(GLOB VISUALIZATION_VISUALIZER_SOURCE_FILES "Visualizer/*.cpp")
 source_group("Source Files\\Visualizer" FILES ${VISUALIZATION_VISUALIZER_SOURCE_FILES})
 
-add_library(Visualization ${Open3D_LIB_TYPE}
+add_library(Visualization 
 	${SHADER_FILES}
 	${VISUALIZATION_ALL_HEADER_FILES} ${VISUALIZATION_ALL_SOURCE_FILES}
 	${VISUALIZATION_UTILITY_HEADER_FILES} ${VISUALIZATION_UTILITY_SOURCE_FILES}
@@ -43,12 +43,12 @@ add_library(Visualization ${Open3D_LIB_TYPE}
 	${VISUALIZATION_VISUALIZER_HEADER_FILES} ${VISUALIZATION_VISUALIZER_SOURCE_FILES}
 )
 add_dependencies(Visualization ShaderFileTarget)
-
+target_link_libraries(Visualization Core IO ${GLFW_LIBRARIES} ${GLEW_LIBRARIES})
 if (Open3D_HEADLESS_RENDERING)
 	find_package(OSMesa REQUIRED)
 	include_directories(${OSMESA_INCLUDE_DIR})
-	target_link_libraries(Visualization Core ${GLFW_LIBRARIES} ${GLEW_LIBRARIES} ${OSMESA_LIBRARY})
+	target_link_libraries(Visualization ${OSMESA_LIBRARY})
 else ()
 	find_package(OpenGL REQUIRED)
-	target_link_libraries(Visualization Core ${GLFW_LIBRARIES} ${GLEW_LIBRARIES} ${OPENGL_LIBRARIES})
+	target_link_libraries(Visualization ${OPENGL_LIBRARIES})
 endif ()
