+++ src/Visualization/Utility/SelectionPolygonVolume.h
@@ -30,7 +30,7 @@
 #include <string>
 #include <memory>
 #include <Eigen/Core>
-#include <IO/ClassIO/IJsonConvertible.h>
+#include <Core/Utility/IJsonConvertible.h>
 
 namespace three {
 
@@ -40,24 +40,24 @@ class PointCloud;
 class SelectionPolygonVolume : public IJsonConvertible
 {
 public:
-	~SelectionPolygonVolume() override {}
-	
+    ~SelectionPolygonVolume() override {}
+
 public:
-	bool ConvertToJsonValue(Json::Value &value) const override;
-	bool ConvertFromJsonValue(const Json::Value &value) override;
-	std::shared_ptr<PointCloud> CropPointCloud(const PointCloud &input) const;
+    bool ConvertToJsonValue(Json::Value &value) const override;
+    bool ConvertFromJsonValue(const Json::Value &value) override;
+    std::shared_ptr<PointCloud> CropPointCloud(const PointCloud &input) const;
 
 private:
-	std::shared_ptr<PointCloud> CropPointCloudInPolygon(
-			const PointCloud &input) const;
-	std::vector<size_t> CropInPolygon(
-			const std::vector<Eigen::Vector3d> &input) const;
+    std::shared_ptr<PointCloud> CropPointCloudInPolygon(
+            const PointCloud &input) const;
+    std::vector<size_t> CropInPolygon(
+            const std::vector<Eigen::Vector3d> &input) const;
 
 public:
-	std::string orthogonal_axis_ = "";
-	std::vector<Eigen::Vector3d> bounding_polygon_;
-	double axis_min_ = 0.0;
-	double axis_max_ = 0.0;
+    std::string orthogonal_axis_ = "";
+    std::vector<Eigen::Vector3d> bounding_polygon_;
+    double axis_min_ = 0.0;
+    double axis_max_ = 0.0;
 };
 
-}	// namespace three
+}   // namespace three
