+++ src/Visualization/Visualizer/ViewParameters.h
@@ -27,34 +27,34 @@
 #pragma once
 
 #include <Eigen/Core>
-#include <IO/ClassIO/IJsonConvertible.h>
+#include <Core/Utility/IJsonConvertible.h>
 
 namespace three {
 
 class ViewParameters : public IJsonConvertible
 {
 public:
-	typedef Eigen::Matrix<double, 17, 4, Eigen::RowMajor> Matrix17x4d;
-	typedef Eigen::Matrix<double, 17, 1> Vector17d;
+    typedef Eigen::Matrix<double, 17, 4, Eigen::RowMajor> Matrix17x4d;
+    typedef Eigen::Matrix<double, 17, 1> Vector17d;
 
 public:
-	ViewParameters() {}
-	~ViewParameters() override {}
+    ViewParameters() {}
+    ~ViewParameters() override {}
 
 public:
-	Vector17d ConvertToVector17d();
-	void ConvertFromVector17d(const Vector17d &v);
-	bool ConvertToJsonValue(Json::Value &value) const override;
-	bool ConvertFromJsonValue(const Json::Value &value) override;
+    Vector17d ConvertToVector17d();
+    void ConvertFromVector17d(const Vector17d &v);
+    bool ConvertToJsonValue(Json::Value &value) const override;
+    bool ConvertFromJsonValue(const Json::Value &value) override;
 
 public:
-	double field_of_view_;
-	double zoom_;
-	Eigen::Vector3d lookat_;
-	Eigen::Vector3d up_;
-	Eigen::Vector3d front_;
-	Eigen::Vector3d boundingbox_min_;
-	Eigen::Vector3d boundingbox_max_;
+    double field_of_view_;
+    double zoom_;
+    Eigen::Vector3d lookat_;
+    Eigen::Vector3d up_;
+    Eigen::Vector3d front_;
+    Eigen::Vector3d boundingbox_min_;
+    Eigen::Vector3d boundingbox_max_;
 };
 
-}	// namespace three
+}   // namespace three
